% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{scrape_nonqna}
\alias{scrape_nonqna}
\title{A Webscraper for collecting data on debates in the Parliament of India}
\usage{
scrape_nonqna(y)
}
\arguments{
\item{y}{Character, String vectors.}
}
\value{
A tibble stored as a vector in the Global Environment.
}
\description{
Collects data from searches on the Parliament of India's digital repository
\url{https://eparlib.nic.in}. When the search is filtered by type "Part 2
Other than Questions and Answers", it returns a data frame with details
including the link to the pdf document in each search result.
}
