% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function1.R
\name{scrape_qna}
\alias{scrape_qna}
\title{A Web Scraper for Collecting Metadata on Questions asked in the Parliament of
India}
\usage{
scrape_qna(x)
}
\arguments{
\item{x}{Character, String vectors.}
}
\value{
A tibble stored as a vector in the Global Environment.
}
\description{
When the search is filtered by type "Part 1(Questions and Answers)", this
function returns a data frame with details including the link to the pdf
document in each search result.
}
