% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function2.R
\name{scrape_nonqna}
\alias{scrape_nonqna}
\title{A Web Scraper for Collecting Metadata on Debates in the Parliament of India}
\usage{
scrape_nonqna(x)
}
\arguments{
\item{x}{Character, String vectors.}
}
\value{
A tibble stored as a vector in the Global Environment.
}
\description{
When the search is filtered by type "Part 2(Other than Questions and
Answers)", this function returns a data frame with details including the link
to the pdf document in each search result.
}
